- removed one warning about double to int conversion reported by user - added function to locate augustus binary path from within augustus
git-svn-id: file:///home/gumz3/Desktop/svnAugustusRepo/svnArchiveFile/svn/augustus/trunks@700 946bced8-b911-4d30-a262-e0a769a8c5ad

Former-commit-id: 4b1a0ebff832e4ef6a20dd8e4719ee4f80d49630

+++ src/exonmodel.cc
@@ -832,9 +832,9 @@ double *ExonModel::getCodonUsage(){
 	pi[s2i(sampledCDS+e.order+i*3)] += 1;
     delete [] sampledCDS;
     // normalize pi
-    int s = 0.0;
+    int s = 0;
     for (int c=0; c<64; c++)
-	s += pi[c];
+	s += (int) pi[c]; // here, the elements of pi are still integers
     
     //    cout << "codon usage: " << endl;
     for (int c=0; c<64; c++){
